update brainordinates/timepoints private variables to cifti return type, which should also fix "total elements" multiplication for large files

+++ src/Files/CiftiConnectivityMatrixDenseDynamicFile.h
@@ -82,9 +82,9 @@ namespace caret {
         
         CiftiFile* m_parentDataSeriesCiftiFile;
         
-        int32_t m_numberOfBrainordinates;
+        int64_t m_numberOfBrainordinates;
         
-        int32_t m_numberOfTimePoints;
+        int64_t m_numberOfTimePoints;
         
         bool m_validDataFlag;
